+++ backend/fittoring/src/main/java/fittoring/application/community/service/PostService.java
@@ -8,7 +8,6 @@ import fittoring.application.community.repository.PostRepository;
 import fittoring.application.community.service.dto.PostCreateDto;
 import fittoring.application.community.service.dto.PostDeleteDto;
 import fittoring.application.community.service.dto.PostPaginationResult;
-import fittoring.application.community.service.dto.PostSearchKeyword;
 import fittoring.application.community.service.dto.PostUpdateDto;
 import fittoring.application.exception.BusinessErrorMessage;
 import fittoring.application.exception.EmptyRequestException;
@@ -16,8 +15,8 @@ import fittoring.application.exception.ForbiddenException;
 import fittoring.application.exception.MemberNotFoundException;
 import fittoring.application.exception.PostNotFoundException;
 import fittoring.application.member.repository.MemberRepository;
-import fittoring.domain.model.Member;
 import fittoring.domain.model.LikeActorKeyHash;
+import fittoring.domain.model.Member;
 import fittoring.domain.model.Post;
 import fittoring.util.CursorCodec;
 import java.util.List;
